pyproject.toml
[tool.poetry]
name = "file-repo"
version = "0.1.0"
description = ""
authors = ["Replit user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
GitPython = "^3.1.14"
uvicorn = "^0.13.4"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
